Address review feedback: use IMAGE_TOKEN_NEWLINE and rename module to unlimited_ocr
- Revert SINGLE_IMAGE_TOKEN_NEWLINE in prompt_utils.py, use IMAGE_TOKEN_NEWLINE
- Rename model directory from unlimitedocr to unlimited_ocr
- Update model type mapping in utils.py to unlimited_ocr
- Update all references in tests and docs
