+++ mlx_vlm/utils.py
@@ -46,7 +46,7 @@ MODEL_REMAPPING = {
     "falcon-perception": "falcon_perception",
     "nemotronh_nano_omni_reasoning_v3": "nemotron_h_nano_omni",
     "cohere2moe": "cohere2_moe",
-    "unlimited-ocr": "unlimitedocr",
+    "unlimited-ocr": "unlimited_ocr",
 }
 
 MAX_FILE_SIZE_GB = 5
